# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Courtpit API",
    "description": "Courtpit REST API.\n\n**Community.** Every request is scoped to one community. Native clients send `X-Courtpit-Community: <slug>`; browsers are resolved from the host (`<slug>.courtpit.app` or a registered custom domain). The header wins over the host. A request without a resolvable community answers `400`, an unknown one `404`.\n\n**Authentication.** A session is an opaque token, sent as `Authorization: Bearer <token>` (native) or as the `courtpit_session` cookie (web; ask for it at sign-in with `X-Courtpit-Client: web`). Either one authenticates.\n\n**Errors.** Every error is `{ \"error\": { \"code\", \"message\" } }`; clients switch on `code`.\n\n**Lists** are cursor-paginated: `?cursor=&limit=` in, `{ items, next_cursor }` out.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/matches/{id}/resolve": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Decides a disputed match (admin).",
        "operationId": "resolve_match",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/admin/matches/{id}/walkover": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Awards an unplayed match to one side (no-show, missed deadline). Admin only.",
        "operationId": "walkover_match",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalkoverBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "The body is malformed or breaks a rule.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/v1/matches/{id}/confirm": {
      "post": {
        "tags": [
          "matches"
        ],
        "summary": "Confirms the other side's reported score. The result then stands.",
        "operationId": "confirm_score",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/matches/{id}/dispute": {
      "post": {
        "tags": [
          "matches"
        ],
        "summary": "Disputes the other side's reported score; a community admin then decides.",
        "operationId": "dispute_score",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "The body is malformed or breaks a rule.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/v1/matches/{id}/score": {
      "post": {
        "tags": [
          "matches"
        ],
        "summary": "Reports the score of a match that was played, whether or not a time was ever agreed\n(`proposed` or `scheduled`); any open proposal is superseded. The other side then has the\ncommunity's confirmation window (default 3 days) to confirm or dispute; silence confirms.",
        "operationId": "report_score",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Score"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Discipline": {
        "type": "string",
        "description": "What kind of tennis is played. Ranked separately (see `mixed_pooling` in scoring).",
        "enum": [
          "singles",
          "doubles",
          "mixed"
        ]
      },
      "DisputeBody": {
        "type": "object",
        "description": "Body of `POST /matches/{id}/dispute`.",
        "properties": {
          "note": {
            "type": [
              "string",
              "null"
            ],
            "description": "What is wrong with the reported score."
          }
        },
        "additionalProperties": false
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The JSON body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "What went wrong."
          }
        }
      },
...
      "MatchFormat": {
        "type": "object",
        "description": "The rules of a match. Stored as JSON on communities, leagues and matches.",
        "required": [
          "sets_to_win",
          "games_per_set",
          "final_set"
        ],
        "properties": {
          "deuce": {
            "$ref": "#/components/schemas/Deuce",
            "description": "How games are decided at deuce. Scores are reported per set, so this does not change\nwhich scores are legal."
          },
          "final_set": {
            "$ref": "#/components/schemas/FinalSet",
            "description": "How the deciding set is played."
          },
          "games_per_set": {
            "type": "integer",
            "format": "int32",
            "description": "Games needed to win a set (usually 6).",
            "minimum": 0
          },
          "sets_to_win": {
            "type": "integer",
            "format": "int32",
            "description": "Sets needed to win: 1, 2 (best of 3) or 3 (best of 5).",
            "minimum": 0
          },
          "tiebreak_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Play a tiebreak at `n–n`. Must equal `games_per_set`; `None` means advantage sets.",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "MatchStatus": {
        "type": "string",
        "description": "Lifecycle status of a match: the flat tag of a [`MatchState`], as stored and serialized.",
        "enum": [
          "proposed",
          "scheduled",
          "reported",
          "confirmed",
          "disputed",
          "resolved",
          "walkover",
          "cancelled"
        ]
      },
      "MatchView": {
        "type": "object",
        "description": "API view of a match.",
        "required": [
          "id",
          "discipline",
          "side_a",
          "side_b",
          "side_a_names",
          "side_b_names",
          "status",
          "match_format",
          "created_at"
        ],
        "properties": {
          "confirm_deadline_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Unanswered reports auto-confirm at this time."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the row was created."
          },
          "discipline": {
            "$ref": "#/components/schemas/Discipline",
            "description": "Singles, doubles or mixed."
          },
          "dispute_note": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why the reported score was disputed."
          },
          "disputed_by": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Player who disputed the reported score."
          },
          "division_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Division the match belongs to, if any."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Match id."
          },
          "league_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "League the match belongs to, if any."
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "description": "Free-text place of play."
          },
          "match_format": {
            "$ref": "#/components/schemas/MatchFormat",
            "description": "Format the match is played in."
          },
          "proposals": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Proposal"
            },
            "description": "Scheduling history; included on single-match responses only."
          },
          "reported_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the score was reported."
          },
          "reported_by": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Player who reported the score."
          },
          "resolution_note": {
            "type": [
              "string",
              "null"
            ],
            "description": "Note left by the resolver."
          },
          "resolved_by": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Player who resolved or cancelled the match."
          },
          "round": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Round within the league or tournament, if any."
          },
          "scheduled_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Agreed start time, if scheduled."
          },
          "score": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Score",
                "description": "Reported score, if any."
              }
            ]
          },
          "side_a": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Player ids on side A."
          },
          "side_a_names": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerName"
            },
            "description": "Side A with display names, in the order of `side_a`."
          },
          "side_b": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Player ids on side B."
          },
          "side_b_names": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerName"
            },
            "description": "Side B with display names, in the order of `side_b`."
          },
          "status": {
            "$ref": "#/components/schemas/MatchStatus",
            "description": "Current lifecycle status."
          },
          "tournament_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Tournament the match belongs to, if any."
          },
          "winner_side": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Side",
                "description": "Winning side, once decided."
              }
            ]
          }
        }
      },
...
      "Resolution": {
        "type": "string",
        "description": "What an admin decides about a disputed result.",
        "enum": [
          "score",
          "replay",
          "void"
        ]
      },
      "ResolveBody": {
        "type": "object",
        "description": "Body of `POST /admin/matches/{id}/resolve`.",
        "required": [
          "resolution"
        ],
        "properties": {
          "note": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional admin note explaining the decision."
          },
          "resolution": {
            "$ref": "#/components/schemas/Resolution",
            "description": "`score`: set the result; `replay`: back to scheduled; `void`: cancel the match."
          },
          "score": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Score",
                "description": "Required with `score`."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "Score": {
        "type": "object",
        "description": "A reported score.",
        "required": [
          "sets"
        ],
        "properties": {
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetScore"
            },
            "description": "The sets in the order they were played."
          }
        },
        "additionalProperties": false
      },
...
      "WalkoverBody": {
        "type": "object",
        "description": "Body of `POST /admin/matches/{id}/walkover`.",
        "required": [
          "winner_side"
        ],
        "properties": {
          "note": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional admin note explaining the walkover."
          },
          "winner_side": {
            "$ref": "#/components/schemas/Side",
            "description": "The side awarded the match."
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token (native clients)."
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "courtpit_session",
        "description": "Session cookie (web clients); alternative to the bearer token."
      }
    }
  },
...
}